pyproject.toml
[tool.poetry]
name = "tandau"
version = "0.1.0"
description = ""
authors = ["Zhangeldi <zhanikshakimurat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
psycopg2-binary = "^2.9.10"
cryptography = "^44.0.2"
openai = "^1.70.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
